requirements.txt
streamlit>=1.28.0
openai>=1.3.0
python-dotenv>=1.0.0
streamlit>=1.28.0
openai>=1.3.0
python-dotenv>=1.0.0
requests>=2.25.0 